feat: removed callout components in markers in order to greatly decrease initial loading time of map
