+++ package.json
@@ -91,6 +91,7 @@
     "lnurl-pay": "^3.0.1",
     "lodash.merge": "^4.6.2",
     "node-libs-react-native": "^1.2.1",
+    "random-location": "^1.1.3",
     "react": "18.2.0",
     "react-content-loader": "^6.2.0",
     "react-dom": "^18.0.0",
